Initial implementation of Scopes
A Scope is like an Optional but with a setter method that always
succeeds, as opposed to the setter method of an Optional that will only
succeed if the getter method succeeds. They are called Scopes (short for
spectroscopes) because they can be built by composing a Lens with a
Prism.

This initial implementation defines the abstract class for a Scope, the
ScopeLaws, enriched methods on Isos, Lenses and Prisms for converting to
Scopes, and methods to compose Isos, Lenses, Prisms, and Scopes in
supported ways. There are property-based tests ensuring that these
methods satisfy the ScopeLaws for some example Scopes formed from Isos,
Lenses and Prisms, along with formal proofs in the comments.