Initial implementation of EPrisms
An EPrism is a Prism with a get method that returns an error message for
values that don't match the Prism. This initial implementation defines
the abstract class for an EPrism, the EPrismLaws, and methods to convert
between Prisms and EPrisms. There are property-based tests ensuring that
these methods satisfy the EPrismLaws for some example EPrisms.