Initial implementation of EScopes
An EScope is a Scope with a get method that returns an error message for
values that don't match the Scope. This initial implementation defines
the abstract class for an EScope, the EScopeLaws, and methods to convert
between Scopes and EScopes. There are property-based tests ensuring that
these methods satisfy the EScopeLaws for some example EScopes.